+++ binaries/coordinator/Cargo.toml
@@ -24,3 +24,4 @@ dora-tracing = { workspace = true, optional = true }
 futures-concurrency = "7.1.0"
 serde_json = "1.0.86"
 names = "0.14.0"
+ctrlc = "3.2.5"
