+++ binaries/daemon/Cargo.toml
@@ -34,3 +34,4 @@ shared-memory-server = { workspace = true }
 bincode = "1.3.3"
 async-trait = "0.1.64"
 aligned-vec = "0.5.0"
+ctrlc = "3.2.5"
